pyproject.toml
[project]
name = "mag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "matplotlib>=3.8.0",
    "tqdm>=4.66.0",
    "tensorboard>=2.16.0",
]
